requirements.txt
asgiref==3.7.2
astroid==3.0.3
backports.zoneinfo==0.2.1
colorama==0.4.6
dill==0.3.8
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
psycopg2==2.9.9
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
